requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scipy>=1.10

# Machine Learning
scikit-learn>=1.3

# Visualization (optional but useful)
matplotlib>=3.7
seaborn>=0.12

# PDF Processing
pikepdf>=8.0
PyMuPDF>=1.23

# Text Extraction (optional fallback tools)
pdfplumber>=0.10
pdfminer.six>=20221105

# OCR
pytesseract>=0.3.10
Pillow>=10.0

# Deep Learning (Image Module)
torch>=2.0
torchvision>=0.15

# Utilities
tqdm>=4.65
python-dotenv>=1.0

# Image Processing (for ELA)
opencv-python>=4.8